HCC:
copy_data.py : cosmetic SCI-HI --> PRIZM changes, added interactive option for selecting which RPis to
copy data from

prizm_daq_2018.py : housecleaning; removed unused functions and old commented out sections, added code
for attempting to read RTC and storing as a separate time field in the data

set_gpstime.py : removed stray "i" from serial.timeout, not sure how this bug didn't cause problems before

+++ copy_data.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env /usr/bin/python
 
 # A dummy script to automatically execute the correct rsync command
-# for copying SCI-HI data.  Looks for existence of external backup
+# for copying PRIZM data.  Looks for existence of external backup
 # drive by default, but also gives an option to copy to the laptop.
 # Checks each of the three Pi IP addresses for connection.
 
@@ -9,12 +9,15 @@ import os
 
 if __name__ == '__main__':
 
+    # We have more than one year now!
+    year = 2018
+    
     # Start by looking for existence of external drive
     try_laptop = False
     extdrive = '/media/scihi/SCIHI_DISK1'
     if os.path.exists(extdrive):
         print 'Detected external drive', extdrive
-        extpath = extdrive+'/marion2017'
+        extpath = extdrive+'/marion'+year
         ret = raw_input('Copy to external drive path '+extpath+'? (y/n) ')
         if ret.lower() == 'y':
             dest = extpath
@@ -25,7 +28,7 @@ if __name__ == '__main__':
         try_laptop = True
     # See if the human wants to copy to the laptop instead
     if try_laptop:
-        extpath = '/data/marion2017'
+        extpath = '/data/marion'+year
         ret = raw_input('Copy to laptop path '+extpath+'? (y/n) ')
         if ret.lower() == 'y':
             dest = extpath
@@ -36,20 +39,36 @@ if __name__ == '__main__':
                 print 'Goodbye...'
                 exit(0)
 
-    # Now try to autodetect which Pi is connected
+    # Now try to autodetect which Pi is connected.  This will work
+    # seamlessly with our new network switch too.
+
+    # Start by asking if the human wants to copy everything or if the
+    # human wants to interactively specify which Pis to copy from.
+    interactive = False
+    ret = raw_input('Copy data from all three RPis (if n, then code will ask interactively)? (y/n)')
+    if ret.lower() == 'n':
+        interactive = True
+        print 'Ok, I will ask you to confirm the data copy from each RPi individually.'
     ip_front = '146.230.92.'
     ip_ends = ['186','187','188']
     data_dir = {'186':'data_100MHz',
                 '187':'data_70MHz',
                 '188':'switch_data'}
     for ip_end in ip_ends:
+        do_copy = True
         ip = ip_front + ip_end
-        print 'Checking for connection to', ip
-        ret = os.system('ping -c 1 -W 2 '+ip)
-        if ret == 0:
-            print 'Found live connection to', ip, ': initializing data transfer'	    
-            cmd = 'rsync -auv --ignore-existing --progress pi@'+ip+':' + data_dir[ip_end] +' '+dest
-            print cmd
-            os.system(cmd)
-        else:
-            print 'Failed to detect connection to', ip
+        if interactive:
+            ret = raw_input('Do you want data from '+ip+'/'+data_dir[ip_end]+'? (y/n)')
+            if ret.lower() == 'n':
+                do_copy = False
+                print 'Your wish is my command -- skipping data copy for '+ip
+        if do_copy:
+            print 'Checking for connection to', ip
+            ret = os.system('ping -c 1 -W 2 '+ip)
+            if ret == 0:
+                print 'Found live connection to', ip, ': initializing data transfer'	    
+                cmd = 'rsync -auv --ignore-existing --progress pi@'+ip+':' + data_dir[ip_end] +' '+dest
+                print cmd
+                os.system(cmd)
+            else:
+                print 'Failed to detect connection to', ip
